Adding test train split script, moving utilities inside directories. Splitting files. Removing JSON for future parse fix
