+++ utilities/all_assertions_from_Crossref_API/requirements.txt
@@ -1,0 +1,5 @@
+certifi==2023.5.7
+charset-normalizer==3.1.0
+idna==3.4
+requests==2.31.0
+urllib3==2.0.3
