+++ utilities/test_train_split/requirements.txt
@@ -1,0 +1,5 @@
+joblib==1.3.1
+numpy==1.25.0
+scikit-learn==1.2.2
+scipy==1.11.1
+threadpoolctl==3.1.0
